+++ bindings/core/src/method_handler/utils.rs
@@ -57,19 +57,29 @@ pub(crate) fn call_utils_method_internal(method: UtilsMethod) -> Result<Response
         UtilsMethod::HashTransactionEssence { essence } => Response::TransactionEssenceHash(prefix_hex::encode(
             TransactionEssence::try_from_dto_unverified(essence)?.hash(),
         )),
-        UtilsMethod::VerifyEd25519Signature {
-            signature,
-            message,
-            address,
-        } => {
-            let signature = Ed25519Signature::try_from(signature)?;
-            let msg: Vec<u8> = prefix_hex::decode(message)?;
-            Response::Bool(signature.is_valid(&msg, &address).is_ok())
-        }
         UtilsMethod::VerifyMnemonic { mut mnemonic } => {
             verify_mnemonic(&mnemonic)?;
             mnemonic.zeroize();
             Response::Ok
+        }
+        UtilsMethod::VerifyEd25519Signature { signature, message } => {
+            let signature = Ed25519Signature::try_from(signature)?;
+            let message: Vec<u8> = prefix_hex::decode(message)?;
+            Response::Bool(signature.verify(&message)?)
+        }
+        UtilsMethod::VerifySecp256k1EcdsaSignature {
+            public_key,
+            signature,
+            message,
+        } => {
+            use crypto::signatures::secp256k1_ecdsa;
+            use iota_sdk::types::block::Error;
+            let public_key = prefix_hex::decode(public_key)?;
+            let public_key = secp256k1_ecdsa::PublicKey::try_from_bytes(&public_key).map_err(Error::from)?;
+            let signature = prefix_hex::decode(signature)?;
+            let signature = secp256k1_ecdsa::Signature::try_from_bytes(&signature).map_err(Error::from)?;
+            let message: Vec<u8> = prefix_hex::decode(message)?;
+            Response::Bool(public_key.verify(&signature, &message))
         }
     };
     Ok(response)